Fix firmware frame buffer bug and rewrite audio plugin
- Fix uint8_t → uint16_t for frame_received bitmask (10 chunks need 10 bits)
- Rewrite audio plugin: use pw-cat with stream.capture.sink for real audio
  capture, rolling window normalization for full 0-100% brightness range,
  48kHz stereo capture at native sink format

+++ firmware/keylume.c
@@ -16,7 +16,7 @@ static bool    frame_dirty = false;
 
 // For SET_FRAME reassembly
 static uint8_t frame_seq       = 0;
-static uint8_t frame_received  = 0;  // bitmask of received chunks (10 chunks max)
+static uint16_t frame_received = 0;  // bitmask of received chunks (10 chunks max)
 static uint16_t frame_expected = 0;   // bitmask for complete frame
 
 static void send_response(uint8_t *data, uint8_t response_code, uint8_t length) {
